+++ src/ConfigElementType/Concrete/ImageGalleryConfigElementType.php
@@ -8,27 +8,16 @@
 
 namespace HeimrichHannot\ConfigElementTypeBundle\ConfigElementType\Concrete;
 
+use Contao\Controller;
 use Contao\FilesModel;
+use Contao\FrontendTemplate;
 use Contao\StringUtil;
-use HeimrichHannot\ConfigElementTypeBundle\ConfigElementType\ConfigElementTypeData;
+use HeimrichHannot\ConfigElementTypeBundle\ConfigElementType\ConfigElementData;
+use HeimrichHannot\ConfigElementTypeBundle\ConfigElementType\ConfigElementResult;
 use HeimrichHannot\ConfigElementTypeBundle\ConfigElementType\ConfigElementTypeInterface;
-use HeimrichHannot\UtilsBundle\Image\ImageUtil;
 
 class ImageGalleryConfigElementType implements ConfigElementTypeInterface
 {
-    /**
-     * @var ImageUtil
-     */
-    protected $imageUtil;
-
-    /**
-     * ImageGalleryConfigElementType constructor.
-     */
-    public function __construct(ImageUtil $imageUtil)
-    {
-        $this->imageUtil = $imageUtil;
-    }
-
     public static function getType(): string
     {
         return 'image_gallery';
@@ -39,23 +28,33 @@ class ImageGalleryConfigElementType implements ConfigElementTypeInterface
         return $prependPalette.'{config_legend},imageSelectorField,imageField,imgSize;'.$appendPalette;
     }
 
-    public function applyConfiguration(ConfigElementTypeData $configElementData): void
+    public function applyConfiguration(ConfigElementData $configElementData): ConfigElementResult
     {
         $configuration = $configElementData->getConfiguration();
-        $item = $configElementData->getItem();
+        $itemData = $configElementData->getItemData();
 
-        if (($configuration->imageSelectorField && $item->getRawValue($configuration->imageSelectorField) && $configuration->imageField && $item->getRawValue($configuration->imageField)) || (!$configuration->imageSelectorField && $configuration->imageField && $item->getRawValue($configuration->imageField))) {
-            $multiSrc = StringUtil::deserialize($item->getRawValue($configuration->imageField));
+        if (($configuration->imageSelectorField && $itemData[$configuration->imageSelectorField] && $configuration->imageField && $itemData[$configuration->imageField]) || (!$configuration->imageSelectorField && $configuration->imageField && $itemData[$configuration->imageField])) {
+            $multiSrc = StringUtil::deserialize($itemData[$configuration->imageField]);
             // Return if there are no files
             if (empty($multiSrc) || !\is_array($multiSrc)) {
-                return;
+                return new ConfigElementResult(ConfigElementResult::TYPE_NONE, null);
             }
 
             // Get the file entries from the database
             $images = FilesModel::findMultipleByUuids($multiSrc);
 
             if (!$images) {
-                return;
+                return new ConfigElementResult(ConfigElementResult::TYPE_NONE, null);
+            }
+
+            $imageSize = null;
+
+            if ($configuration->imgSize) {
+                $size = StringUtil::deserialize($configuration->imgSize);
+
+                if ($size[0] > 0 || $size[1] > 0 || is_numeric($size[2])) {
+                    $imageSize = $configuration->imgSize;
+                }
             }
 
             $galleryData = [];
@@ -63,17 +62,15 @@ class ImageGalleryConfigElementType implements ConfigElementTypeInterface
             foreach ($images as $index => $filesModel) {
                 $imageArray['singleSRC'] = $filesModel->path;
 
-                if ($configuration->imgSize) {
-                    $size = StringUtil::deserialize($configuration->imgSize);
-
-                    if ($size[0] > 0 || $size[1] > 0 || is_numeric($size[2])) {
-                        $imageArray['size'] = $configuration->imgSize;
-                    }
+                if ($imageSize) {
+                    $imageArray['size'] = $imageSize;
                 }
-                $galleryData[$index] = [];
-                $this->imageUtil->addToTemplateData('singleSRC', '', $galleryData[$index], $imageArray, null, null, null, $filesModel);
+                $template = new FrontendTemplate();
+                Controller::addImageToTemplate($template, $imageArray);
+                $galleryData[] = $template->getData();
             }
         }
-        $item->setFormattedValue($configuration->templateVariable, $galleryData);
+
+        return new ConfigElementResult(ConfigElementResult::TYPE_FORMATTED_VALUE, $galleryData);
     }
 }