+++ src/Resources/contao/languages/en/tl_reader_config_element.php
@@ -6,8 +6,6 @@
  * @license LGPL-3.0-or-later
  */
 
-$GLOBALS['TL_LANG']['tl_reader_config_element']['reference'][\HeimrichHannot\ConfigElementTypeBundle\ConfigElementType\Concrete\ImageGalleryConfigElementType::getType()] = 'Bildergallerie';
-
 $lang = &$GLOBALS['TL_LANG']['tl_reader_config_element'];
 
 $lang['reference'][\HeimrichHannot\ConfigElementTypeBundle\ConfigElementType\Concrete\ImageGalleryConfigElementType::getType()] = 'Image gallery';